CV2-6018: rename folder to be SavedSearches instead of Projects
