+++ localization/react-intl/src/app/components/drawer/SavedSearches/NewSavedSearch.json
@@ -1,0 +1,17 @@
+[
+  {
+    "id": "projectsComponent.placeholder",
+    "description": "Placeholder for creating a new custom list",
+    "defaultMessage": "Enter a short, easily remembered name for this custom list"
+  },
+  {
+    "id": "projectsComponent.title",
+    "description": "Text field label for the title input",
+    "defaultMessage": "Title"
+  },
+  {
+    "id": "NewSavedSearch.cancel",
+    "description": "Dialog label for the cancel button to close the dialog",
+    "defaultMessage": "Cancel"
+  }
+]
