+++ localization/react-intl/src/app/components/drawer/SavedSearches/SavedSearchActions.json
@@ -1,0 +1,67 @@
+[
+  {
+    "id": "SavedSearchActions.tooltip",
+    "description": "Toolitp for the button that shows actions that can be performed on a list",
+    "defaultMessage": "Actions"
+  },
+  {
+    "id": "SavedSearchActions.defaultErrorMessage",
+    "description": "Generic error message displayed when it's not possible to update or delete a list",
+    "defaultMessage": "Error, please try again"
+  },
+  {
+    "id": "SavedSearchActions.savedSuccessfully",
+    "description": "Generic success message displayed when a list is updated or deleted",
+    "defaultMessage": "Done"
+  },
+  {
+    "id": "SavedSearchActions.rename",
+    "description": "'Rename' here is an infinitive verb",
+    "defaultMessage": "Rename"
+  },
+  {
+    "id": "SavedSearchActions.delete",
+    "description": "'Delete' here is an infinitive verb",
+    "defaultMessage": "Delete"
+  },
+  {
+    "id": "projectsComponent.title",
+    "description": "Label for the title input when renaming a list",
+    "defaultMessage": "Title"
+  },
+  {
+    "id": "global.cancel",
+    "description": "Generic label for a button or link for a user to press when they wish to abort an in-progress operation",
+    "defaultMessage": "Cancel"
+  },
+  {
+    "id": "projectsComponent.renameType",
+    "description": "'Rename' here is an infinitive verb",
+    "defaultMessage": "Rename list"
+  },
+  {
+    "id": "savedSearchActions.deleteMessageWarning",
+    "description": "A message that appears when a user tries to delete a list, warning them that it will affect other users in their workspace.",
+    "defaultMessage": "Are you sure? This is shared among all users of <strong>{teamName}</strong>. After deleting it, no user will be able to access it.<br /><br />"
+  },
+  {
+    "id": "saveFeed.deleteCustomListWarning",
+    "description": "Warning displayed on edit feed page when no list is selected.",
+    "defaultMessage": "<strong>Deleting list will result in no content for the following shared feeds:</strong>"
+  },
+  {
+    "id": "savedSearchActions.deleteMessage",
+    "description": "A message that appears when a user tries to delete a list, warning them that it will affect other users in their workspace.",
+    "defaultMessage": "Are you sure? This is shared among all users of <strong>{teamName}</strong>. After deleting it, no user will be able to access it."
+  },
+  {
+    "id": "projectsComponent.deleteType",
+    "description": "'Delete' here is an infinitive verb",
+    "defaultMessage": "Delete list"
+  },
+  {
+    "id": "projectsComponent.deleteTitleType",
+    "description": "'Delete' here is an infinitive verb",
+    "defaultMessage": "Delete list?"
+  }
+]
