+++ src/app/components/drawer/DrawerFeeds.js
@@ -5,13 +5,13 @@ import PropTypes from 'prop-types';
 import { FormattedMessage, FormattedHTMLMessage, injectIntl, defineMessages } from 'react-intl';
 import { browserHistory, withRouter } from 'react-router';
 import cx from 'classnames/bind';
-import SavedSearchesListItem from './Projects/SavedSearchesListItem';
+import SavedSearchesListItem from './SavedSearches/SavedSearchesListItem';
 import ButtonMain from '../cds/buttons-checkboxes-chips/ButtonMain';
 import AddIcon from '../../icons/add.svg';
 import Can from '../Can';
 import ScheduleSendIcon from '../../icons/schedule_send.svg';
 import { withSetFlashMessage } from '../FlashMessage';
-import styles from './Projects/SavedSearches.module.css';
+import styles from './SavedSearches/SavedSearches.module.css';
 
 const messages = defineMessages({
   pendingInvitationFeedTooltip: {
