+++ src/app/components/drawer/SavedSearches/DrawerListCounter.js
