+++ src/app/components/drawer/SavedSearches/DrawerListCounter.test.js
